+++ src/Server/Process/Foreground.php
@@ -15,7 +15,10 @@ use Innmind\Immutable\{
     Predicate\Instance,
 };
 
-final class Foreground implements Process
+/**
+ * @internal
+ */
+final class Foreground
 {
     private Started $process;
     /** @var Sequence<Chunk> */
@@ -23,7 +26,7 @@ final class Foreground implements Process
     /** @var ?Either<Failed|Signaled|TimedOut, Success> */
     private ?Either $status = null;
 
-    public function __construct(Started $process, bool $streamOutput = false)
+    public function __construct(Started $process, bool $streamOutput)
     {
         $this->process = $process;
         $yieldOutput = function() use ($process): \Generator {
@@ -58,19 +61,25 @@ final class Foreground implements Process
         $this->output = $output;
     }
 
-    #[\Override]
+    /**
+     * @return Maybe<Pid>
+     */
     public function pid(): Maybe
     {
         return Maybe::of($this->process->pid());
     }
 
-    #[\Override]
+    /**
+     * @return Sequence<Chunk>
+     */
     public function output(): Sequence
     {
         return $this->output;
     }
 
-    #[\Override]
+    /**
+     * @return Either<TimedOut|Failed|Signaled, Success>
+     */
     public function wait(): Either
     {
         if (\is_null($this->status)) {