+++ tests/Servers/LoggerTest.php
@@ -7,10 +7,16 @@ use Innmind\Server\Control\{
     Servers\Logger,
     Server,
     Server\Processes,
+    Server\Processes\Unix,
     Server\Process,
+    Server\Process\Pid,
     Server\Command,
     Server\Volumes,
+    Server\Signal,
 };
+use Innmind\TimeContinuum\Earth\Clock;
+use Innmind\TimeWarp\Halt\Usleep;
+use Innmind\Stream\Streams;
 use Innmind\Immutable\{
     Either,
     SideEffect,
@@ -28,25 +34,15 @@ class LoggerTest extends TestCase
         $this->assertInstanceOf(
             Server::class,
             Logger::psr(
-                $this->createMock(Server::class),
-                $this->createMock(LoggerInterface::class),
+                $this->server(),
+                new NullLogger,
             ),
         );
     }
 
     public function testProcesses()
     {
-        $server = $this->createMock(Server::class);
-        $server
-            ->expects($this->once())
-            ->method('processes')
-            ->willReturn($processes = $this->createMock(Processes::class));
-        $processes
-            ->expects($this->once())
-            ->method('execute')
-            ->with($this->callback(static function(Command $command): bool {
-                return $command->toString() === 'ls';
-            }));
+        $server = $this->server('ls');
 
         $logger = Logger::psr(
             $server,
@@ -62,41 +58,7 @@ class LoggerTest extends TestCase
 
     public function testVolumes()
     {
-        $server = $this->createMock(Server::class);
-        $server
-            ->expects($this->once())
-            ->method('processes')
-            ->willReturn($processes = $this->createMock(Processes::class));
-        $which1 = $this->createMock(Process::class);
-        $which1
-            ->expects($this->once())
-            ->method('wait')
-            ->willReturn(Either::right(new SideEffect));
-        $which2 = $this->createMock(Process::class);
-        $which2
-            ->expects($this->once())
-            ->method('wait')
-            ->willReturn(Either::right(new SideEffect));
-        $processes
-            ->expects($matcher = $this->exactly(2))
-            ->method('execute')
-            ->willReturnCallback(function(Command $command) use ($matcher, $which1, $which2) {
-                match ($matcher->numberOfInvocations()) {
-                    1 => $this->assertSame(
-                        'which diskutil',
-                        $command->toString(),
-                    ),
-                    2 => $this->assertSame(
-                        "diskutil 'unmount' '/dev'",
-                        $command->toString(),
-                    ),
-                };
-
-                return match ($matcher->numberOfInvocations()) {
-                    1 => $which1,
-                    2 => $which2,
-                };
-            });
+        $server = $this->server('which diskutil', "diskutil 'unmount' '/dev'");
 
         $logger = Logger::psr(
             $server,
@@ -127,22 +89,7 @@ class LoggerTest extends TestCase
 
     public function testReboot()
     {
-        $server = $this->createMock(Server::class);
-        $server
-            ->expects($this->once())
-            ->method('processes')
-            ->willReturn($processes = $this->createMock(Processes::class));
-        $processes
-            ->expects($this->once())
-            ->method('execute')
-            ->with($this->callback(static function(Command $command): bool {
-                return $command->toString() === 'sudo shutdown -r now';
-            }))
-            ->willReturn($shutdown = $this->createMock(Process::class));
-        $shutdown
-            ->expects($this->once())
-            ->method('wait')
-            ->willReturn(Either::right(new SideEffect));
+        $server = $this->server('sudo shutdown -r now');
 
         $logger = Logger::psr(
             $server,
@@ -170,22 +117,7 @@ class LoggerTest extends TestCase
 
     public function testShutdown()
     {
-        $server = $this->createMock(Server::class);
-        $server
-            ->expects($this->once())
-            ->method('processes')
-            ->willReturn($processes = $this->createMock(Processes::class));
-        $processes
-            ->expects($this->once())
-            ->method('execute')
-            ->with($this->callback(static function(Command $command): bool {
-                return $command->toString() === 'sudo shutdown -h now';
-            }))
-            ->willReturn($shutdown = $this->createMock(Process::class));
-        $shutdown
-            ->expects($this->once())
-            ->method('wait')
-            ->willReturn(Either::right(new SideEffect));
+        $server = $this->server('sudo shutdown -h now');
 
         $logger = Logger::psr(
             $server,
@@ -208,6 +140,69 @@ class LoggerTest extends TestCase
                 static fn($sideEffect) => $sideEffect,
                 static fn() => null,
             ),
+        );
+    }
+
+    private function server(string ...$commands): Server
+    {
+        return new class($this->processes(), $this, $commands) implements Server {
+            private $inner;
+
+            public function __construct(
+                private $processes,
+                private $test,
+                private $commands,
+            ) {
+            }
+
+            public function processes(): Processes
+            {
+                return $this->inner ??= new class($this->processes, $this->test, $this->commands) implements Processes {
+                    public function __construct(
+                        private $processes,
+                        private $test,
+                        private $commands,
+                    ) {
+                    }
+
+                    public function execute(Command $command): Process
+                    {
+                        $expected = \array_shift($this->commands);
+                        $this->test->assertNotNull($expected);
+                        $this->test->assertSame(
+                            $expected,
+                            $command->toString(),
+                        );
+
+                        return $this->processes->execute(Command::foreground('echo'));
+                    }
+
+                    public function kill(Pid $pid, Signal $signal): Either
+                    {
+                    }
+                };
+            }
+
+            public function volumes(): Volumes
+            {
+            }
+
+            public function reboot(): Either
+            {
+            }
+
+            public function shutdown(): Either
+            {
+            }
+        };
+    }
+
+    private function processes(): Unix
+    {
+        return Unix::of(
+            new Clock,
+            Streams::fromAmbientAuthority(),
+            new Usleep,
         );
     }
 }